rewrite frame_protocol.py using yield from
The main thing in this commit is a rewrite of the frame_protocol.py
state machine as straight-through linear code, using 'yield
from'. This creates a hard requirement for Python 3.3+. But given that
no-one uses 3.3, and that 3.3 is missing enum, and that 3.3's
bytearray has quadratic behavior in the way I'm using it, in practice
I think the minimum supported version should be 3.4.

I think the code is much, much clearer.

Behaviorally, the main change is that we know emit partial data frames
as we receive them. This is (almost) invisible to the client (it just
looks like peer is sending smaller fragments -- except that we do
expose where the actual fragment boundaries were in case there are any
experts who really want to know this), while avoiding potential DoS
issues and allowing users to get better streaming behavior. This also
flushed out a bunch of bugs related to *sending* fragments, which I
guess was never exercised before.

I believe (but haven't checked) that it should be faster too -- in
particular from removing the str/bytes += loop in add_message_payload.

It also gives better error messages on protocol errors.

It also fixes some technically unrelated bugs I noticed along the way:

- Lots of fragmentation-related bugs in PerMessageDeflate code
- Replace some O(n**2) data structure usage with O(n) equivalents
- Report correct CLOSE codes in more places
- Fix spelling of TLS_HANDSHAKE_FAILED
- Don't echo local-only CLOSE reasons back to the server
- Maybe some others I'm forgetting...