Rewrite to eliminate yield froms.
The theory in this rewrite is to try and break the protocol into some layers.
The layers in question are the frame decoder, the message decoder and the
frame protocol.

The frame decoder takes bytes and turns them into frames. A frame will either
be an entire control message or all or part of a larger text or binary message.

The message decoder handles state that handles message state across frames. It
replaces the CONTINUATION opcode with the relevant TEXT or BINARY opcode and
handles UTF8 decoding for TEXT messages.

The frame protocol deals with upper layer protocol issues such as close value
checking.