Rework our Travis matrix and add pypy.
This keeps us to one run per Python by default, with wsaccel installed when
it makes sense, i.e. CPython uses wsaccel and PyPy does not.

We also add a couple of pytest-only runs with wsaccel not installed to make
sure that our fallbacks work under CPython.

Autobahn is not run under PyPy.