Bugfix Connection header acceptance
Some clients, for example Firefox, will send "Connection: keep-alive,
Upgrade" rather than just "Connection: Upgrade". As the previous
version expected the header value to be exactly Upgrade it would raise
a protocol error of "Missing Connection: Upgrade header". This fixes
the issue by accepting any collection of values in the Connection
header that includes "Upgrade".

The relevant part of rfc6455 seems to be in section 4.1:

    The request MUST contain a |Connection| header field whose value
    MUST include the "Upgrade" token.