Update to h11 0.8.1 and fix test regression
h11 0.8.1 does not allow empty header values anymore so when the
header value is empty we only add the header name without a ; and
whitespace.

This was originally written by @DaanDeMeyer, however that repo has
been deleted and the commit reference lost. This is a restoration, bar
for a typo fix and changelog entry.