Add the ability to initiate an upgrade connection
This is useful if the connection is being used as a server, but
something else has parsed the initial request. This is quite common
for dual protocol (HTTP & Websocket) servers whereby the HTTP part
parses first.

The implementation reverses the parsing, which is a little unpleasent,
however it is more unpleasent to put the h11 upgrade_connection into a
specific state.