Rename and restructure the events
This changes the event naming so that they are all verb baseds and
hence can either be used to represent something that has happened (as
before) or to represent something that should happen. The latter
enables a ``send(event: Event)`` API similar to h11 and h2 to be
created.

Structure wise the h11 EventBundle system (and in places some of the
code) is adopted. This structure is roughly a stripped down
``dataclass``, and when wsproto supports 3.7 or greater I recommend
switching to a dataclass implementation. This restructure standardises
the reprs.

When migrating to this commit be sure to change names as follows,

    ConnectionRequested -> Request
    ConnectionEstablished -> AcceptConnection
    ConnectionClosed -> CloseConnection
    ConnectionFailed -> Fail
    DataReceived -> Data
    TextReceived -> TextMessage
    BytesReceived -> BytesMessage
    PingReceived -> Ping
    PongReceived -> Pong

Note the attributes (fields) may have to change in a subsequent
commit, so this should not be considered a reference commit.