Introduce a send method to the Connection class
This allows a new event based API to be used in preference to the
previous method based API. The basic concept being that callingg
``connection.foo()`` becomes ``connection.send(Foo())``. This follows
the h11 convention.

This requires the following changes to be made to existing code,
(assuming connection is a Connection instance),

    connection.accept(subprotocol=subprotocol) -> connection.send(AcceptConnection(subprotocol=subprotocol))
    connection.send_data(data) -> connection.send(Data(payload=payload))
    connection.close(code) -> connection.send(CloseConnection(code=code))
    connection.ping() -> connection.send(Ping())
    connection.pong() -> connection.send(Pong())