Add functionality for Websocket connection rejection
This allows for a SERVER-WSConnection to reject the websocket upgrade
request with a custom response (including a body), it also allows for
a CLIENT-WSConnection to emit events relating to a response that has
rejected the upgrade request.

The event API is written to short cut the most likely path in the
server case, i.e. rejecting with a 400 code and no body. This is
opposed to say copying the h11 event API.