Expose connection state as a read only property
With the rejection functionality (and likely for other cases) it is
necessary to know the current state of the connection. This was
partially exposed before via the ``closed`` property. However
properties for each state is less efficient that exposing the
state. The state is read only as it should only be changed by sent or
received events.