Switch to LocalProtocolErrors
This is a more pythonic way of understanding if the local action is
awry. In this case it means that either the action has been attempted
in the wrong state or the values do not conform the the websocket
standard. The naming and structure follows that introduced and used by
h11.