Add an event hint to the RemoteProtocolError
This hints to the user of wsproto to how they should respond to the
client based on the client's error. This is particularly useful for
the case whereby a client sends a version we don't support. In this
case a 426 with a specific header is the expected response, which
without a hint seems tricky to remember.