Seperate the handshake code from the connection code
This only changes current usage in that `WSConnection` must be
imported from `wsproto` rather than `wsproto.connection` and that
`receive_bytes` is renamed to `receive_data`. The latter is to be
consistent with the naming and match h11 (as with other API changes).

This allows users to complete the Hanshake themselves (aside from
wsproto) and/or use only the open connection. It will allow for a
HTTP/2 handshake to be developed.