Ensure RemoteProtocolErrors are raised from the receive_data method
This really helps users of wsproto handle the error cases as they are
raised from known actions either send for Local or receive for
Remote. It also ensures that an error is raised, rather than a
CloseConnection event being emitted (which suggests the connection
received a close frame). This is equivalent to the changes to the Fail
event only for the remote case.