Support HTTP/2 WebSockets
This only includes helper functions for the handshake phase. This is
because the HTTP/2 connection must be managed alongside the chosen I/O
(in order for the flow control to work). The recommendation is to use
the Hyper-h2 library with wsproto.

In order to support h2 client extension handshakes the
handshake_extensions function has been renamed.