Drop support for Python 3.5, min version now 3.6.1
This allows for dataclasses to be used (backport for 3.6 exists, but
not 3.5) it also allows for the usage of PEP 526 variable annotations
in the upcoming typing commit.

3.6.1 Introduces typing for Deque which will be required to type hint
the codebase.

Sadly PyPy CI has had to be removed as Travis has no PyPy3.6 support
yet (although the testing did very little due to performance concerns,
see tox.ini).