Switch to aiohttp masking method
This utilises the masking method used in aiohttp as it has better
performance, in addition the aiohttp test is also used. With thanks to
@willmcgugan for the original implementation.

This also removes the option of using wsaccel, it was previously
usable as an undocumented optional extra. This is considered
acceptable as the aiohttp masking method has similar performance to
wsaccel.