Prevent invalid window bit sizes
This is specifically to deal with the case whereby the window size is
set to 8. As zlib dropped support for 256 (8 bit) window sizes in
1.2.9, https://www.zlib.net/ChangeLog.txt, onwards. 8 bit windows were
broken before as well, https://github.com/madler/zlib/issues/171. As
autobahn tests 8 bit windows the tox test would fail based on the zlib
version installed - hence the relevant autobahn tests are also
disabled.

8 bit windows are rarely used in practice, so this should be ok
practically. If not zlib needs to be fixed, see
https://github.com/madler/zlib/issues/87#issuecomment-389333346