Make the value of the upgrade header the same as the RFC.
As per https://datatracker.ietf.org/doc/html/rfc6455#section-4.2.1 the "Upgrade" header should have
the value "websocket", and the server should treat it as case-insensitive.

However, sadly some servers don't respect that case-insentitivity, so better to use exactly the
string as specified in the RFC.

That's what the other websocket implementations do. See for example:
https://github.com/python-websockets/websockets/blob/94dd203f63bb52b1a30faa228e63ada2f0f2e874/src/websockets/client.py#L119